% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leo.R
\name{leo}
\alias{leo}
\title{Print the name Leo}
\usage{
leo()
}
\value{
A \code{character(1)} with the words \code{Leo}.
}
\description{
This function is a test function that returns the name
Leo.
}
\examples{
## Print the name Leo
leo()

## Capitalize the first letter
gsub("l", "L", leo())
}
